backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
cramjam==2.8.3
Django==5.0.7
django-cors-headers==4.7.0
dnspython==2.6.1
geographiclib==2.0
geopy==2.4.1
joblib==1.4.2
numpy==2.1.3
pandas==2.2.3
pillow==11.1.0
PyJWT==2.8.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-snappy==0.7.1
pytz==2024.1
reportlab==4.3.1
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
winkerberos==0.10.0
XlsxWriter==3.2.2